pyproject.toml
[tool.poetry]
name = "mctl"
version = "0.1.0"
description = "mctl"
authors = ["naps-dev"]
readme = "README.md"
packages = [{include = "cli", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
openpyxl = "^3.0.10"
click = "^8.1.3"
toml = "^0.10.2"
python-libmaas = {git = "https://github.com/maas/python-libmaas.git"}
requests = "^2.28.2"
requests-oauthlib = "^1.3.1"
oauthlib = "^3.2.2"
requests-toolbelt = "^0.10.1"
yaspin = "^2.3.0"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.9"
mypy = "^0.991"
flake8-for-pycharm = "^0.4.1"
autoflake = "^2.0.0"
pytest = "^7.2.1"
types-openpyxl = "^3.0.4.4"
ruff = "^0.0.252"

[tool.ruff]
ignore = [
    "E501",
    "E731"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mctl = 'cli:main.cli'
